Allow specifying wn-client location. (SOFTWARE-1144)
The default wn-client location used to be under /net/nas01/Public/ouser.mis/osg-wn-client/...
This was the staging area for the wn-client that we put into CVMFS.
We can't use any of the tools in that directory though because they are looking
for files in the final location for that wn-client
(/cvmfs/oasis.opensciencegrid.org/osg-software/...), which doesn't exist on
oasis-login.opensciencegrid.org (which is where we are expecting this script to
run). Instead, I just made a tarball-client install in ouser.mis's home
directory and am specifying that as the default. I also made it a command-line
option so we can change it later.

Since we can specify the wn-client location explicitly, I no longer need code
that looks at the architecture and distro version to find the wn-client
location. Because of that, I could remove the 'platform' module dependency as
well.

git-svn-id: file:///p/vdt/workspace/svn/software/tarball-client/trunk@17579 4e558342-562e-0410-864c-e07659590f8c