Remove empty dir trees from the tarballs (SOFTWARE-1769)
I extract them, use find to remove the directories and tar them up
again. This is ugly, but the only other option is to determine which
directories will be empty from the exclude list when creating the
tarballs in the first place and that would be difficult to implement.

There is a hack to keep var/lib/osg-ca-certs and etc/fetch-crl.d, since
we do want those in the tarballs.